product connections moved to frontend-api package
