+++ packages/frontend-api/src/Model/Resolver/Products/ProductsQuery.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Shopsys\FrontendApiBundle\Model\Resolver\Products;
 
 use GraphQL\Executor\Promise\Promise;
-use InvalidArgumentException;
+use GraphQL\Type\Definition\ResolveInfo;
 use Overblog\DataLoader\DataLoaderInterface;
 use Overblog\GraphQLBundle\Definition\Argument;
 use Ramsey\Uuid\Uuid;
@@ -14,15 +14,20 @@ use Shopsys\FrameworkBundle\Model\CategorySeo\ReadyCategorySeoMix;
 use Shopsys\FrameworkBundle\Model\Product\Brand\Brand;
 use Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterData;
 use Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterDataFactory;
+use Shopsys\FrameworkBundle\Model\Product\Flag\Flag;
 use Shopsys\FrameworkBundle\Model\Product\List\ProductList;
 use Shopsys\FrameworkBundle\Model\Product\List\ProductListFacade;
 use Shopsys\FrameworkBundle\Model\Product\ProductRepository;
 use Shopsys\FrontendApiBundle\Component\Validation\PageSizeValidator;
 use Shopsys\FrontendApiBundle\Model\Product\BatchLoad\ProductBatchLoadByEntityData;
+use Shopsys\FrontendApiBundle\Model\Product\Connection\ProductConnection;
 use Shopsys\FrontendApiBundle\Model\Product\Connection\ProductConnectionFactory;
 use Shopsys\FrontendApiBundle\Model\Product\Filter\ProductFilterFacade;
 use Shopsys\FrontendApiBundle\Model\Product\ProductFacade;
 use Shopsys\FrontendApiBundle\Model\Resolver\AbstractQuery;
+use Shopsys\FrontendApiBundle\Model\Resolver\Brand\BrandQuery;
+use Shopsys\FrontendApiBundle\Model\Resolver\Category\CategoryQuery;
+use Shopsys\FrontendApiBundle\Model\Resolver\Flag\FlagQuery;
 
 class ProductsQuery extends AbstractQuery
 {
@@ -36,6 +41,9 @@ class ProductsQuery extends AbstractQuery
      * @param \Shopsys\FrontendApiBundle\Model\Resolver\Products\ProductOrderingModeProvider $productOrderingModeProvider
      * @param \Shopsys\FrameworkBundle\Model\Product\Filter\ProductFilterDataFactory $productFilterDataFactory
      * @param \Overblog\DataLoader\DataLoaderInterface $productsByEntitiesBatchLoader
+     * @param \Shopsys\FrontendApiBundle\Model\Resolver\Brand\BrandQuery $brandQuery
+     * @param \Shopsys\FrontendApiBundle\Model\Resolver\Category\CategoryQuery $categoryQuery
+     * @param \Shopsys\FrontendApiBundle\Model\Resolver\Flag\FlagQuery $flagQuery
      */
     public function __construct(
         protected readonly ProductFacade $productFacade,
@@ -47,47 +55,20 @@ class ProductsQuery extends AbstractQuery
         protected readonly ProductOrderingModeProvider $productOrderingModeProvider,
         protected readonly ProductFilterDataFactory $productFilterDataFactory,
         protected readonly DataLoaderInterface $productsByEntitiesBatchLoader,
+        protected readonly BrandQuery $brandQuery,
+        protected readonly CategoryQuery $categoryQuery,
+        protected readonly FlagQuery $flagQuery,
     ) {
     }
 
     /**
      * @param \Overblog\GraphQLBundle\Definition\Argument $argument
-     * @return \Overblog\GraphQLBundle\Relay\Connection\ConnectionInterface|object
-     */
-    public function productsQuery(Argument $argument)
-    {
-        $search = $argument['search'] ?? '';
-
-        $this->setDefaultFirstOffsetIfNecessary($argument);
-
-        $productFilterData = $this->productFilterFacade->getValidatedProductFilterDataForAll(
-            $argument,
-        );
-
-        return $this->productConnectionFactory->createConnectionForAll(
-            function ($offset, $limit) use ($argument, $productFilterData, $search) {
-                return $this->productFacade->getFilteredProductsOnCurrentDomain(
-                    $limit,
-                    $offset,
-                    $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
-                    $productFilterData,
-                    $search,
-                );
-            },
-            $this->productFacade->getFilteredProductsCountOnCurrentDomain($productFilterData, $search),
-            $argument,
-            $productFilterData,
-        );
-    }
-
-    /**
-     * @param \Overblog\GraphQLBundle\Definition\Argument $argument
      * @param \Shopsys\FrameworkBundle\Model\Product\Brand\Brand $brand
-     * @return \Overblog\GraphQLBundle\Relay\Connection\ConnectionInterface|object
+     * @return \GraphQL\Executor\Promise\Promise
      */
-    public function productsByBrandQuery(Argument $argument, Brand $brand)
+    public function productsByBrandQuery(Argument $argument, Brand $brand): Promise
     {
-        $search = $argument['search'] ?? '';
+        PageSizeValidator::checkMaxPageSize($argument);
 
         $this->setDefaultFirstOffsetIfNecessary($argument);
 
@@ -95,22 +76,28 @@ class ProductsQuery extends AbstractQuery
             $argument,
             $brand,
         );
+        $batchLoadDataId = Uuid::uuid4()->toString();
 
-        return $this->productConnectionFactory->createConnectionForBrand(
+        return $this->productConnectionFactory->createConnectionPromiseForBrand(
             $brand,
-            function ($offset, $limit) use ($argument, $brand, $productFilterData, $search) {
-                return $this->productFacade->getFilteredProductsByBrand(
-                    $brand,
-                    $limit,
-                    $offset,
-                    $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
-                    $productFilterData,
-                    $search,
+            function ($offset, $limit) use ($argument, $productFilterData, $brand, $batchLoadDataId) {
+                return $this->productsByEntitiesBatchLoader->load(
+                    new ProductBatchLoadByEntityData(
+                        $batchLoadDataId,
+                        $brand->getId(),
+                        Brand::class,
+                        $limit,
+                        $offset,
+                        $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
+                        $productFilterData,
+                    ),
                 );
             },
-            $this->productFacade->getFilteredProductsByBrandCount($brand, $productFilterData, $search),
             $argument,
             $productFilterData,
+            $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
+            $this->productOrderingModeProvider->getDefaultOrderingMode($argument),
+            $batchLoadDataId,
         );
     }
 
@@ -156,7 +143,7 @@ class ProductsQuery extends AbstractQuery
             );
             $orderingMode = $this->productOrderingModeProvider->getOrderingModeFromArgument($argument);
             $defaultOrderingMode = $this->productOrderingModeProvider->getDefaultOrderingMode($argument);
-        } elseif ($categoryOrReadyCategorySeoMix instanceof ReadyCategorySeoMix) {
+        } else {
             $category = $categoryOrReadyCategorySeoMix->getCategory();
             $readyCategorySeoMix = $categoryOrReadyCategorySeoMix;
 
@@ -169,14 +156,6 @@ class ProductsQuery extends AbstractQuery
 
             $orderingMode = $categoryOrReadyCategorySeoMix->getOrdering();
             $defaultOrderingMode = $orderingMode;
-        } else {
-            throw new InvalidArgumentException(
-                sprintf(
-                    'The "$categoryOrReadyCategorySeoMix" argument must be an instance of "%s" or "%s".',
-                    Category::class,
-                    ReadyCategorySeoMix::class,
-                ),
-            );
         }
 
         return $this->getPromiseByCategory($argument, $category, $productFilterData, $orderingMode, $defaultOrderingMode, $readyCategorySeoMix);
@@ -224,6 +203,106 @@ class ProductsQuery extends AbstractQuery
             $defaultOrderingMode,
             $batchLoadDataId,
             $readyCategorySeoMix,
+        );
+    }
+
+    /**
+     * @param \Overblog\GraphQLBundle\Definition\Argument $argument
+     * @param \Shopsys\FrameworkBundle\Model\Product\Flag\Flag $flag
+     * @return \GraphQL\Executor\Promise\Promise
+     */
+    public function productsByFlagQuery(Argument $argument, Flag $flag): Promise
+    {
+        PageSizeValidator::checkMaxPageSize($argument);
+
+        $this->setDefaultFirstOffsetIfNecessary($argument);
+
+        $productFilterData = $this->productFilterFacade->getValidatedProductFilterDataForFlag(
+            $argument,
+            $flag,
+        );
+
+        $productFilterData->flags[] = $flag;
+        $batchLoadDataId = Uuid::uuid4()->toString();
+
+        return $this->productConnectionFactory->createConnectionPromiseForFlag(
+            $flag,
+            function ($offset, $limit) use ($argument, $productFilterData, $flag, $batchLoadDataId) {
+                return $this->productsByEntitiesBatchLoader->load(
+                    new ProductBatchLoadByEntityData(
+                        $batchLoadDataId,
+                        $flag->getId(),
+                        Flag::class,
+                        $limit,
+                        $offset,
+                        $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
+                        $productFilterData,
+                    ),
+                );
+            },
+            $argument,
+            $productFilterData,
+            $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
+            $this->productOrderingModeProvider->getDefaultOrderingMode($argument),
+            $batchLoadDataId,
+        );
+    }
+
+    /**
+     * @param \Overblog\GraphQLBundle\Definition\Argument $argument
+     * @param \GraphQL\Type\Definition\ResolveInfo $info
+     * @return \Shopsys\FrontendApiBundle\Model\Product\Connection\ProductConnection|\GraphQL\Executor\Promise\Promise
+     */
+    public function productsWithOverlyingEntityQuery(Argument $argument, ResolveInfo $info): ProductConnection|Promise
+    {
+        PageSizeValidator::checkMaxPageSize($argument);
+
+        $this->setDefaultFirstOffsetIfNecessary($argument);
+
+        if ($argument['categorySlug'] !== null) {
+            $category = $this->categoryQuery->categoryOrSeoMixByUuidOrUrlSlugQuery($info, null, $argument['categorySlug']);
+
+            return $this->productsByCategoryOrReadyCategorySeoMixQuery(
+                $argument,
+                $category,
+            );
+        }
+
+        if ($argument['brandSlug'] !== null) {
+            $brand = $this->brandQuery->brandByUuidOrUrlSlugQuery(null, $argument['brandSlug']);
+
+            return $this->productsByBrandQuery(
+                $argument,
+                $brand,
+            );
+        }
+
+        if ($argument['flagSlug'] !== null) {
+            $flag = $this->flagQuery->flagByUuidOrUrlSlugQuery(null, $argument['flagSlug']);
+
+            return $this->productsByFlagQuery(
+                $argument,
+                $flag,
+            );
+        }
+
+        $productFilterData = $this->productFilterFacade->getValidatedProductFilterDataForAll(
+            $argument,
+        );
+
+        return $this->productConnectionFactory->createConnectionForAll(
+            function ($offset, $limit) use ($argument, $productFilterData) {
+                return $this->productFacade->getFilteredProductsOnCurrentDomain(
+                    $limit,
+                    $offset,
+                    $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
+                    $productFilterData,
+                );
+            },
+            $this->productFacade->getFilteredProductsCountOnCurrentDomain($productFilterData),
+            $argument,
+            $productFilterData,
+            $this->productOrderingModeProvider->getOrderingModeFromArgument($argument),
         );
     }
 }
