+++ project-base/app/config/services_frontend_api.yaml
@@ -36,11 +36,6 @@ services:
             $brandsBatchLoader: '@brands_batch_loader'
             $productsSellableCountByIdsBatchLoader: '@products_sellable_count_by_ids_batch_loader'
 
-    App\FrontendApi\Resolver\Products\ProductsQuery:
-        arguments:
-            $productsVisibleAndSortedByIdsBatchLoader: '@products_visible_and_sorted_by_ids_batch_loader'
-            $productsByEntitiesBatchLoader: '@products_by_entities_batch_loader'
-
     App\FrontendApi\Resolver\Store\StoreAvailabilityResolverMap:
         arguments:
             $storesBatchLoader: '@stores_batch_loader'
@@ -78,9 +73,6 @@ services:
     Shopsys\FrontendApiBundle\Model\Parameter\ParameterWithValuesFactory:
         alias: App\FrontendApi\Model\Parameter\ParameterWithValuesFactory
 
-    Shopsys\FrontendApiBundle\Model\Product\Connection\ProductConnectionFactory:
-        alias: App\FrontendApi\Model\Product\Connection\ProductConnectionFactory
-
     Shopsys\FrontendApiBundle\Model\Product\Filter\ProductFilterFacade:
         alias: App\FrontendApi\Model\Product\Filter\ProductFilterFacade
 
@@ -111,7 +103,3 @@ services:
 
     Shopsys\FrontendApiBundle\Model\Resolver\Products\ProductResolverMap:
         alias: App\FrontendApi\Resolver\Products\ProductResolverMap
-
-    Shopsys\FrontendApiBundle\Model\Resolver\Products\ProductsQuery:
-        alias: App\FrontendApi\Resolver\Products\ProductsQuery
-        public: true
